+++ src/lib/client/graphql/get_open_prs.rs
@@ -1,5 +1,6 @@
 #![allow(dead_code)]
 use serde::{Deserialize, Serialize};
+use tracing::instrument;
 
 use crate::{Client, Error, GraphQLWrapper};
 
@@ -59,8 +60,9 @@ pub(crate) struct PrItem {
 }
 
 impl GraphQLGetOpenPRs for Client {
+    #[instrument(skip(self))]
     async fn get_open_pull_requests(&self) -> Result<Vec<PrItem>, Error> {
-        log::trace!("get_open_pull_requests");
+        tracing::trace!("get_open_pull_requests");
         let query = r#"
         query($owner:String!, $name:String!){
             repository(owner: $owner, name: $name) {
@@ -87,11 +89,11 @@ impl GraphQLGetOpenPRs for Client {
             .query_with_vars_unwrap::<Data, Vars>(query, vars)
             .await;
 
-        log::trace!("data_res: {:?}", data_res);
+        tracing::trace!("data_res: {:?}", data_res);
 
         let data = data_res.map_err(GraphQLWrapper::from)?;
 
-        log::trace!("data: {:?}", data);
+        tracing::trace!("data: {:?}", data);
 
         let edges = data
             .repository